broke up editor.rs into dedicated module
